backend/requirements.txt
fastapi
uvicorn
gitpython
pydantic
openai
networkx
python-dotenv
httpx
google-generativeai
